Split out shared networking code into net_traits crate
Fixes #4476
