+++ components/compositing/constellation.rs
@@ -20,9 +20,10 @@ use msg::constellation_msg::{IFrameSandboxState, MozBrowserEvent, NavigationDire
 use msg::constellation_msg::{Key, KeyState, KeyModifiers, LoadData};
 use msg::constellation_msg::{SubpageId, WindowSizeData};
 use msg::constellation_msg::{self, ConstellationChan, Failure};
-use net::image_cache_task::{ImageCacheTask, ImageCacheTaskClient};
-use net::resource_task::{self, ResourceTask};
-use net::storage_task::{StorageTask, StorageTaskMsg};
+use net::image_cache_task::ImageCacheTaskClient;
+use net_traits::{self, ResourceTask};
+use net_traits::image_cache_task::ImageCacheTask;
+use net_traits::storage_task::{StorageTask, StorageTaskMsg};
 use profile::mem;
 use profile::time;
 use script_traits::{CompositorEvent, ConstellationControlMsg};
@@ -392,7 +393,7 @@ impl<LTF: LayoutTaskFactory, STF: ScriptTaskFactory> Constellation<LTF, STF> {
             pipeline.exit(PipelineExitType::Complete);
         }
         self.image_cache_task.exit();
-        self.resource_task.send(resource_task::ControlMsg::Exit).unwrap();
+        self.resource_task.send(net_traits::ControlMsg::Exit).unwrap();
         self.devtools_chan.as_ref().map(|chan| {
             chan.send(DevtoolsControlMsg::ServerExitMsg).unwrap();
         });
