+++ components/compositing/pipeline.rs
@@ -16,11 +16,11 @@ use gfx::font_cache_task::FontCacheTask;
 use layers::geometry::DevicePixel;
 use msg::constellation_msg::{ConstellationChan, Failure, FrameId, PipelineId, SubpageId};
 use msg::constellation_msg::{LoadData, WindowSizeData, PipelineExitType, MozBrowserEvent};
-use net::image_cache_task::ImageCacheTask;
-use net::resource_task::ResourceTask;
-use net::storage_task::StorageTask;
 use profile::mem;
 use profile::time;
+use net_traits::ResourceTask;
+use net_traits::image_cache_task::ImageCacheTask;
+use net_traits::storage_task::StorageTask;
 use std::sync::mpsc::{Receiver, channel};
 use url::Url;
 use util::geometry::{PagePx, ViewportPx};
