+++ components/gfx/Cargo.toml
@@ -11,8 +11,8 @@ path = "lib.rs"
 [dependencies.plugins]
 path = "../plugins"
 
-[dependencies.net]
-path = "../net"
+[dependencies.net_traits]
+path = "../net_traits"
 
 [dependencies.util]
 path = "../util"
