+++ components/gfx/lib.rs
@@ -31,7 +31,7 @@ extern crate profile;
 extern crate script_traits;
 extern crate "rustc-serialize" as rustc_serialize;
 extern crate unicode;
-extern crate net;
+extern crate net_traits;
 #[macro_use]
 extern crate util;
 extern crate msg;
