+++ components/layout/context.rs
@@ -13,9 +13,9 @@ use gfx::display_list::OpaqueNode;
 use gfx::font_cache_task::FontCacheTask;
 use gfx::font_context::FontContext;
 use msg::constellation_msg::ConstellationChan;
-use net::local_image_cache::LocalImageCache;
 use script::layout_interface::{Animation, LayoutChan};
 use script_traits::UntrustedNodeAddress;
+use net_traits::local_image_cache::LocalImageCache;
 use std::boxed;
 use std::cell::Cell;
 use std::ptr;
