+++ components/layout/fragment.rs
@@ -30,8 +30,8 @@ use gfx::text::text_run::{TextRun, TextRunSlice};
 use script_traits::UntrustedNodeAddress;
 use rustc_serialize::{Encodable, Encoder};
 use msg::constellation_msg::{ConstellationChan, Msg, PipelineId, SubpageId};
-use net::image::holder::ImageHolder;
-use net::local_image_cache::LocalImageCache;
+use net_traits::image::holder::ImageHolder;
+use net_traits::local_image_cache::LocalImageCache;
 use util::geometry::{self, Au, ZERO_POINT};
 use util::logical_geometry::{LogicalRect, LogicalSize, LogicalMargin, WritingMode};
 use util::range::*;
