+++ components/layout/layout_task.rs
@@ -41,12 +41,12 @@ use log;
 use msg::compositor_msg::ScrollPolicy;
 use msg::constellation_msg::Msg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Failure, PipelineExitType, PipelineId};
-use net::image_cache_task::{ImageCacheTask, ImageResponseMsg};
-use net::local_image_cache::{ImageResponder, LocalImageCache};
-use net::resource_task::{ResourceTask, load_bytes_iter};
 use profile::mem::{self, Report, ReportsChan};
 use profile::time::{self, ProfilerMetadata, profile};
 use profile::time::{TimerMetadataFrameType, TimerMetadataReflowType};
+use net_traits::{load_bytes_iter, ResourceTask};
+use net_traits::image_cache_task::{ImageCacheTask, ImageResponseMsg};
+use net_traits::local_image_cache::{ImageResponder, LocalImageCache};
 use script::dom::bindings::js::LayoutJS;
 use script::dom::node::{LayoutData, Node};
 use script::layout_interface::{Animation, ContentBoxResponse, ContentBoxesResponse};
