+++ components/layout_traits/Cargo.toml
@@ -16,8 +16,8 @@ path = "../script_traits"
 [dependencies.msg]
 path = "../msg"
 
-[dependencies.net]
-path = "../net"
+[dependencies.net_traits]
+path = "../net_traits"
 
 [dependencies.profile]
 path = "../profile"
