+++ components/net/file_loader.rs
@@ -2,8 +2,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use resource_task::{ProgressMsg, Metadata, LoadData, start_sending, TargetedLoadResponse, ResponseSenders};
-use resource_task::ProgressMsg::{Payload, Done};
+use net_traits::{LoadData, Metadata, ProgressMsg};
+use net_traits::ProgressMsg::{Payload, Done};
+use resource_task::{start_sending, TargetedLoadResponse, ResponseSenders};
 
 use std::borrow::ToOwned;
 use std::io;
