+++ components/net_traits/image/holder.rs
