+++ components/script/dom/bindings/trace.rs
@@ -48,8 +48,8 @@ use js::rust::{Cx, rt};
 use layout_interface::{LayoutRPC, LayoutChan};
 use libc;
 use msg::constellation_msg::{PipelineId, SubpageId, WindowSizeData, WorkerId};
-use net::image_cache_task::ImageCacheTask;
-use net::storage_task::StorageType;
+use net_traits::image_cache_task::ImageCacheTask;
+use net_traits::storage_task::StorageType;
 use script_traits::ScriptControlChan;
 use script_traits::UntrustedNodeAddress;
 use msg::compositor_msg::ScriptListener;
