+++ components/script/dom/document.rs
@@ -63,8 +63,8 @@ use msg::compositor_msg::ScriptListener;
 use msg::constellation_msg::Msg as ConstellationMsg;
 use msg::constellation_msg::{ConstellationChan, Key, KeyState, KeyModifiers, MozBrowserEvent};
 use msg::constellation_msg::{SUPER, ALT, SHIFT, CONTROL};
-use net::resource_task::ControlMsg::{SetCookiesForUrl, GetCookiesForUrl};
-use net::cookie_storage::CookieSource::NonHTTP;
+use net_traits::CookieSource::NonHTTP;
+use net_traits::ControlMsg::{SetCookiesForUrl, GetCookiesForUrl};
 use script_task::Runnable;
 use script_traits::{MouseButton, UntrustedNodeAddress};
 use util::{opts, namespace};
