+++ components/script/lib.rs
@@ -39,7 +39,7 @@ extern crate hyper;
 extern crate js;
 extern crate libc;
 extern crate msg;
-extern crate net;
+extern crate net_traits;
 extern crate "rustc-serialize" as rustc_serialize;
 extern crate time;
 extern crate canvas;
