+++ components/script/script_task.rs
@@ -60,10 +60,10 @@ use msg::constellation_msg::{ConstellationChan};
 use msg::constellation_msg::{LoadData, PipelineId, SubpageId, MozBrowserEvent, WorkerId};
 use msg::constellation_msg::{Failure, WindowSizeData, PipelineExitType};
 use msg::constellation_msg::Msg as ConstellationMsg;
-use net::image_cache_task::ImageCacheTask;
-use net::resource_task::{ResourceTask, ControlMsg, LoadResponse};
-use net::resource_task::LoadData as NetLoadData;
-use net::storage_task::StorageTask;
+use net_traits::{ResourceTask, ControlMsg, LoadResponse};
+use net_traits::LoadData as NetLoadData;
+use net_traits::image_cache_task::ImageCacheTask;
+use net_traits::storage_task::StorageTask;
 use string_cache::Atom;
 use util::geometry::to_frac_px;
 use util::smallvec::SmallVec;
