+++ components/script_traits/Cargo.toml
@@ -10,8 +10,8 @@ path = "lib.rs"
 [dependencies.msg]
 path = "../msg"
 
-[dependencies.net]
-path = "../net"
+[dependencies.net_traits]
+path = "../net_traits"
 
 [dependencies.util]
 path = "../util"
