+++ components/servo/Cargo.toml
@@ -48,6 +48,9 @@ path = "../compositing"
 [dependencies.net]
 path = "../net"
 
+[dependencies.net_traits]
+path = "../net_traits"
+
 [dependencies.msg]
 path = "../msg"
